requirements.txt
fastapi
uvicorn
python-multipart
pyzbar
opencv-python
validators
pillow
